Remove custom hook for stripe url
